DOC-188 modified Tests for DocumentFQN and increased Coverage Level
