+++ datasafe-simple-adapter/datasafe-simple-adapter-api/pom.xml
@@ -22,6 +22,13 @@
             <artifactId>slf4j-api</artifactId>
             <version>${slf4j-simple.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.junit.jupiter</groupId>
+            <artifactId>junit-jupiter-api</artifactId>
+            <version>5.4.1</version>
+            <scope>test</scope>
+        </dependency>
+
 
     </dependencies>
