+++ datasafe-simple-adapter/datasafe-simple-adapter-api/src/test/java/de/adorsys/datasafe/simple/adapter/api/DocumentDirectoryFQNTest.java
@@ -1,4 +1,4 @@
-package de.adorsys.datasafe.simple.adapter.impl;
+package de.adorsys.datasafe.simple.adapter.api;
 
 
 import de.adorsys.datasafe.simple.adapter.api.types.DocumentDirectoryFQN;
@@ -17,25 +17,14 @@ public class DocumentDirectoryFQNTest {
         Assertions.assertEquals("/a/b/c", d.getDocusafePath());
     }
     @Test
-    public void slashesForFiles() {
-        DocumentFQN d = new DocumentFQN("/a/b/c");
-        Assertions.assertEquals("/a/b/c", d.getDocusafePath());
-        d = new DocumentFQN("/a/b/c/");
-        Assertions.assertEquals("/a/b/c", d.getDocusafePath());
-        d = new DocumentFQN("/a///b//c/");
-        Assertions.assertEquals("/a/b/c", d.getDocusafePath());
-    }
-
-    @Test
-    public void startingSlashForDocumentFQN() {
-        DocumentFQN d = new DocumentFQN("/a/b/c");
-        Assertions.assertEquals("/a/b/c", d.getDocusafePath());
-        Assertions.assertEquals("a/b/c", d.getDatasafePath());
-    }
-    @Test
     public void startingSlashForDocumentDirectoryFQN() {
         DocumentDirectoryFQN d = new DocumentDirectoryFQN("/a/b/c");
         Assertions.assertEquals("/a/b/c", d.getDocusafePath());
         Assertions.assertEquals("a/b/c", d.getDatasafePath());
+    }
+    @Test void emptyDocumentDirectoryFQN() {
+        DocumentDirectoryFQN d = new DocumentDirectoryFQN("");
+        Assertions.assertEquals("/", d.getDocusafePath());
+        Assertions.assertEquals("", d.getDatasafePath());
     }
 }
